login kadai

+++ app/Http/Controllers/TasksController.php
@@ -15,24 +15,20 @@ class TasksController extends Controller
      */
     public function index()
     {
-        // $tasks = Task::all();
         
           $data = [];
         if (\Auth::check()) {
             $user = \Auth::user();
             $tasks = $user->tasks()->orderBy('created_at', 'desc')->paginate(10);
             
-            $data = [
+           $data = [
                 'user' => $user,
                 'tasks' => $tasks,
             ];
         }
         
-        return view('welcome', $data);
+        return view('tasks.index' , $data);
         
-       // return view('tasks.index', [
-        //    'tasks' => $tasks,
-        //]);
     }
 
     /**
@@ -62,10 +58,11 @@ class TasksController extends Controller
             'content' => 'required|max:191',
         ]);
         
-        $task = new Task;
-        $task->status = $request->status;
-        $task->content = $request->content;
-        $task->save();
+       $request->user()->tasks()->create([
+            'content' => $request->content,
+            'status' => $request->status,
+        ]);
+        
         
         return redirect('/');
     }
@@ -130,9 +127,11 @@ class TasksController extends Controller
      */
     public function destroy($id)
     {
-        $task = Task::find($id);
-        $task->delete();
         
+        $task = Task::find($id);
+         if (\Auth::id() === $task->user_id) {
+        $task->delete();
+         }
         return redirect('/');
     }
 }
